Factor out example vs examples
This can be compressed by making a mutually exclusve "oneOf"
with the proper definitions of each property, where each branch
of the "oneOf" also forbids the other property with a {"not": {}}
schema.

This allows defining all objects that use examples with

    example: {}
    examples: {}

just to exempt the names from the additionalProperties: false,
while the real validation of these fields is handled by
using a "allOf" to overlay the "oneOf" onto the objects.