Verbose Header Object documentation (3.1.1)
This copies the relevant Parameter Object fields to the Header Object
instead of relying on implicit guidance.  The text for the fields
has been edited to reflect that only headers are being described.

This also include an example of describing a header using the
`content` field, and explaining why it is necessary to do so.