Explain Param and Header example serialization
The rules for this have not been clear, and are not always intuitive.
This states and explains them directly and ensures that the
Style Examples table matches the rules.

Unlike past efforts, this provides a rule system regarding what
is and is not included, based on a combination of what is produced
by RFC6570 (or the nearest RFC6570 equivalent), modified by
removing leading delimiters that are not correct for our usage
due to differences from the assuptions made by RFC6570.

This also shows some uses of the new Example Object fields,
including some that would be redundant but are included to
clarify the different options; the redundancy is noted in the text.